Fix issue where TransitionPresence would transitions on parent rerender
The TransitionPresence receives "new" children when the parent component
rerenders for whichever reason. More often than not the same component
(JSX) is passed, so there is no need for a in/out animation.

This update expects all children of the TransitionPresence to have a
`key` prop defined, and uses that to determine if a new child is passed
and if it should trigger the beforeUnmount.

This update also includes the `onStart` and `onComplete` handlers on the
TransitionPresence to trigger when transitions start and finish.