Improve setting animations by validating the input
If nullish keys are passed to be used in the AnimationMap, we cannot
reliably store and retrieve these values, since it's most likely an
error somewhere up.

When setting values that haven't changed since the last render, we make
sure to not trigger any listeners, as nothing would have changed.