Improve rerender checks in TransitionPresence
The switch to `memo` didn't work out since it would block too much
rendering when updating other props (or children) of the children passed
to the TransitionPresence.

A story is added for the above scenario, so we won't make the same
mistake in the future.

Some comments and logging is added since this component will receive
other updates shortly.